Changed how the event system works
Basicallt now the program itself would run inside a coroutine, every time this main coroutine yields, the kernel will block everything and try to get an input. this allows for the program to create its own coroutines and catch when one of its coroutine yields and process that yield however it wants, the idea is to make so the program will be able to manage multiple coroutines in a "parallel" way.