added parallel API
the parallel API allows you to run multiple coroutines until either one, or all of them finish. the way it works is by running each of the coroutines until a yield (assuming from os_pull_event) and then pull a single event and resume each of the coroutines with that event.